requirements.txt
# Core — required for all usage
requests[socks]>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
stem>=1.8.2
lxml>=5.0.0

# LLM providers — install only what you use
openai>=1.30.0
anthropic>=0.28.0
google-generativeai>=0.7.0

# LangChain / CrewAI integration (optional)
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langchain-ollama>=0.1.0

# MCP server (optional — needed only for: python sicry.py serve)
mcp>=1.0.0
